+++ src/app/utils/dependencies.py
@@ -2,6 +2,7 @@
 
 # pylint: disable=missing-function-docstring
 from fastapi.security import OAuth2PasswordBearer
+from passlib.context import CryptContext
 
 from app.database.settings import local_session
 
@@ -15,3 +16,5 @@ def database_session():
 
 
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
+
+password_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
